ProF/back/requirements.txt
Django==3.2
pyJWT==2.1.0
cloudinary
django-cors-headers
django-environ
djangorestframework
djoser
social-auth-app-django

django-ckeditor

pytz
Pillow
argon2-cffi
django-environ
stripe
braintree
boto3
django-storages
markdown
gunicorn
whitenoise
requests

mutagen

# ///

# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
# asgiref==3.7.2
# boto3==1.28.62
# botocore==1.31.62
# braintree==4.22.0
# certifi==2023.7.22
# cffi==1.16.0
# charset-normalizer==3.3.0
# cryptography==41.0.4
# defusedxml==0.8.0rc2
# Django==4.2
# django-ckeditor==6.7.0
# django-cors-headers==4.2.0
# django-environ==0.11.2
# django-js-asset==2.1.0
# django-storages==1.14.1
# django-templated-mail==1.1.1
# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.3.0
# djoser==2.2.0
# gunicorn==21.2.0
# idna==3.4
# jmespath==1.0.1
# Markdown==3.5
# oauthlib==3.2.2
# packaging==23.2
# Pillow==10.0.1
# pycparser==2.21
# PyJWT==2.8.0
# python-dateutil==2.8.2
# python3-openid==3.2.0
# pytz==2023.3.post1
# requests==2.31.0
# requests-oauthlib==1.3.1
# s3transfer==0.7.0
# six==1.16.0
# social-auth-app-django==5.3.0
# social-auth-core==4.4.2
# sqlparse==0.4.4
# stripe==6.7.0
# typing_extensions==4.8.0
# tzdata==2023.3
# urllib3==2.0.6
# whitenoise==6.5.0